% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kagomer-package.R
\docType{package}
\name{kagomer-package}
\alias{kagomer}
\alias{kagomer-package}
\title{kagomer: Asynchronous HTTP Client for Kagome Server}
\description{
An asynchronous HTTP client for the Kagome server
    powered by gaborcsardi/async.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/kagomer}
  \item Report bugs at \url{https://github.com/paithiov909/kagomer/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
