% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{kick}
\alias{kick}
\title{Kicks a request}
\usage{
kick(
  request,
  keep = c("surface", "class", "base_form", "reading"),
  .skip_enc_reset = FALSE
)
}
\arguments{
\item{request}{An asynchronous function to be kicked.}

\item{keep}{Column names to keep in the results.}

\item{.skip_enc_reset}{Logical. If true, leaves encoding of results as the environment native.}
}
\value{
data.frame.
}
\description{
Kicks a request
}
