% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Sends a HEAD request to Kagome server}
\usage{
ping(url = "http://localhost:6060")
}
\arguments{
\item{url}{URL stirng.}
}
\value{
returns the status code of response invisibly.
}
\description{
Sends a HEAD request to Kagome server
}
