% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{serialize}
\alias{serialize}
\title{Creates json data}
\usage{
serialize(sentence, mode = c("normal"), dict = c("ipa", "uni"))
}
\arguments{
\item{sentence}{Characters to be analyzed.}

\item{mode}{One of `normal`, `search` or `extended`.}

\item{dict}{One of `ipa` (the IPA-dic) or `uni` (the Uni-dic).}
}
\value{
returns a json string.
}
\description{
Creates json data
}
