% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{kick}
\alias{kick}
\title{Kicks requests}
\usage{
kick(
  requests,
  keep = c("surface", "class", "base_form", "reading"),
  .skip_enc_reset = FALSE
)
}
\arguments{
\item{keep}{Column names to keep in the results.}

\item{.skip_enc_reset}{Logical. If true, leaves encoding of results as environment native.}

\item{request}{List that comes out from \code{queue}.}
}
\value{
data.frame.
}
\description{
Kicks requests
}
