% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping.R
\name{ping}
\alias{ping}
\title{Send a HEAD request to Kagome server}
\usage{
ping(url = "http://localhost:6060")
}
\arguments{
\item{url}{URL stirng.}
}
\value{
the status code of response is returned invisibly.
}
\description{
Send a HEAD request to Kagome server
}
