% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{queue}
\alias{queue}
\title{Creates an asynchronous request}
\usage{
queue(params, url = "http://localhost:6060/tokenize")
}
\arguments{
\item{params}{List of json string.}

\item{url}{An URL of Kagome server API.}
}
\value{
returns a list of json string and asynchronous function.
}
\description{
Creates an asynchronous request
}
