% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{queue}
\alias{queue}
\title{Creates an asynchronous request}
\usage{
queue(params, url = "http://localhost:6060/tokenize")
}
\arguments{
\item{url}{URL of Kagome server API.}
}
\value{
returns a asynchronous function.
}
\description{
Creates an asynchronous request
}
