% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{serialize}
\alias{serialize}
\title{Creates json data}
\usage{
serialize(
  sentences,
  mode = c("normal", "search", "extended"),
  dict = c("ipa", "uni")
)
}
\arguments{
\item{mode}{One of `normal`, `search` or `extended`.}

\item{dict}{One of `ipa` (the IPA-dic) or `uni` (the Uni-dic).}

\item{sentence}{Character to be analyzed.}
}
\value{
returns a json string.
}
\description{
Creates json data
}
